requirements_local.txt
asgiref==3.8.1
dj-database-url==2.3.0
dj-static==0.0.6
Django==5.2
django-cors-headers==4.7.0
django-heroku==0.3.1
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==24.2
psycopg2==2.9.10
python-decouple==3.8
sqlparse==0.5.3
static3==0.7.0
typing_extensions==4.13.2
tzdata==2025.2
whitenoise==6.9.0
